+++ Packages/com.unity.inputsystem/InputSystem/Devices/IFlushable.cs
@@ -1,0 +1,7 @@
+namespace UnityEngine.InputSystem.LowLevel
+{
+    internal interface IFlushable
+    {
+        void Flush();
+    }
+}