+++ Packages/com.unity.inputsystem/InputSystem/InputManager.cs
@@ -2854,6 +2854,12 @@ namespace UnityEngine.InputSystem
                 m_InputEventStream.Advance(leaveEventInBuffer: leaveInBuffer);
             }
 
+            foreach (var device in devices)
+            {
+                if (device is IFlushable flushable)
+                    flushable.Flush();
+            }
+
             m_Metrics.totalEventProcessingTime += ((double)(Stopwatch.GetTimestamp() - processingStartTime)) / Stopwatch.Frequency;
             m_Metrics.totalEventLagTime += totalEventLag;